requirements.txt
jieba
whoosh
django-haystack
Django==2.2.5
django-appconf==1.0.4
django-ckeditor==5.9.0
django-imagekit==4.0.2
django-mdeditor==0.1.18
django-notifications-hq==1.6.0
django-simpleui==4.0
djangorestframework==3.11.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
Pillow==7.1.2
PyMySQL==0.9.3
requests==2.23.0
sqlparse==0.3.1
swapper==1.1.2.post1
uritemplate==3.0.1
urllib3==1.25.9
zipp==3.1.0
